Fix issues with using date library under ruby 1.9.3
